introduce check_interval in addition to run_time
Currently the proper full check of the refcount and state variables
can be done after all the threads are finished. This is because doing it
during the run would require locking and if the locking is broken (which
this tool is supposed to test) it's best to avoid it.

Introduce check_interval in seconds after which the processes are
restarted so that the full check can be performed while still respecting
the specified run_time.

Set the default check interval to random number of seconds between 1 and
61 and the default runtime to 10 minutes.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>